Refactor - improve readability of code and make more functional
I've extracted out the prompt methods to their own file as
there will be quite a few of them and it will make the main cli.ts
file look bloated and hard to understand.

By having the bookSelector method as the entry point for the cli,
it becomes more clear what the outline of the logic is.

I've also separated the methods, so that they are not nested within
one another. This pattern will be beneficial when the cli requests even
more input, otherwise it would be a lot of nested questions and it would
start to get messy. Each method now returns a promise that the next method can
wait for before executing.